+++ src/util/numeric.h
@@ -165,7 +165,7 @@ inline float modulo360f(float f)
 		sign = 1;
 	}
 
-	whole = f;
+	whole = (int)f;
 
 	fraction = f - whole;
 	whole %= 360;
@@ -272,9 +272,9 @@ inline constexpr f32 sqr(f32 f)
 inline v3s16 floatToInt(v3f p, f32 d)
 {
 	return v3s16(
-		(p.X + (p.X > 0 ? d / 2 : -d / 2)) / d,
-		(p.Y + (p.Y > 0 ? d / 2 : -d / 2)) / d,
-		(p.Z + (p.Z > 0 ? d / 2 : -d / 2)) / d);
+		    (s16)((p.X + (p.X > 0 ? d / 2 : -d / 2)) / d),
+			(s16)((p.Y + (p.Y > 0 ? d / 2 : -d / 2)) / d),
+			(s16)((p.Z + (p.Z > 0 ? d / 2 : -d / 2)) / d));
 }
 
 /*
@@ -282,10 +282,9 @@ inline v3s16 floatToInt(v3f p, f32 d)
  */
 inline v3s16 doubleToInt(v3d p, double d)
 {
-	return v3s16(
-		(p.X + (p.X > 0 ? d / 2 : -d / 2)) / d,
-		(p.Y + (p.Y > 0 ? d / 2 : -d / 2)) / d,
-		(p.Z + (p.Z > 0 ? d / 2 : -d / 2)) / d);
+	return v3s16((s16)((p.X + (p.X > 0 ? d / 2 : -d / 2)) / d),
+			(s16)((p.Y + (p.Y > 0 ? d / 2 : -d / 2)) / d),
+			(s16)((p.Z + (p.Z > 0 ? d / 2 : -d / 2)) / d));
 }
 
 /*
